Upgrade of SPDX version
SPDX version upgraded to the 3.17 version.
After the upgrade the licensedb.Detect() function begun to
panic with:
panic: runtime error: index out of range [0] with length 0.
The root cause was in the queryLicenseAbstract method of the
database. After adding a length check to amount of matches
found the issue disappeared.
The upgrade also affected a confidence of the matched licenses
in the TestCmdMain, so they were adjusted accordingly to pass.