Get license full name and license URLs
In order to extend functionality the two new exportable functions
were added to get more info about the license.
LicenseURLs returns a list of URLs for the license and
LicenseName returns a full name of the license. Both accepts a
license identifier as an input parameter and do search over the
license database. If there is no info about provided license the
error ErrUnknownLicenseID is returned.
The extract_urls.go was extended to produce one more column in
urls.csv file that includes schema of the URL. It is needed to
construct the full URL that will be returned from the LicenseURLs.